+++ src/app.ts
@@ -1,12 +1,12 @@
 import FastifyCORS from 'fastify-cors';
 import FastifySensible from 'fastify-sensible';
-import { IS_DEV } from 'common/config';
-import { buildApp } from 'common/build-app';
-import { prettifier } from 'common/logger';
-import { setupOpenApiDoc } from 'openaip/setup';
-import { userRouter } from 'resources/users/user.router';
-import { boardRouter } from 'resources/boards/board.router';
-import { taskRouter } from 'resources/tasks/task.router';
+import { buildApp } from '~src/common/build-app';
+import { IS_DEV } from '~src/common/config';
+import { prettifier } from '~src/common/logger';
+import { setupOpenApiDoc } from '~src/openaip/setup';
+import { boardRouter } from '~src/resources/boards/board.router';
+import { taskRouter } from '~src/resources/tasks/task.router';
+import { userRouter } from '~src/resources/users/user.router';
 
 export const app = buildApp({}, IS_DEV && prettifier);
