+++ src/resources/boards/board.memory-repository.ts
@@ -1,4 +1,4 @@
-import { IDBRepository, MemoryRepository } from 'common/memory-repository';
+import { IDBRepository, MemoryRepository } from '~src/common/memory-repository';
 import { Board } from './board.model';
 
 export const boardsRepo: IDBRepository<Board> = new MemoryRepository();
