+++ src/resources/tasks/task.memory.repository.ts
@@ -1,4 +1,4 @@
-import { IDBRepository, MemoryRepository } from 'common/memory-repository';
+import { IDBRepository, MemoryRepository } from '~src/common/memory-repository';
 import { Task } from './task.model';
 
 export const tasksRepo: IDBRepository<Task> = new MemoryRepository();
