+++ src/resources/users/user.memory-repository.ts
@@ -1,4 +1,4 @@
-import { IDBRepository, MemoryRepository } from 'common/memory-repository';
+import { IDBRepository, MemoryRepository } from '~src/common/memory-repository';
 import { User } from './user.model';
 
 export const usersRepo: IDBRepository<User> = new MemoryRepository();
