+++ src/server.ts
@@ -1,5 +1,5 @@
-import { PORT } from 'common/config';
-import { app } from 'app';
+import { PORT } from '~src/common/config';
+import { app } from './app';
 
 app.listen(PORT, (err) => {
   app.log.info(`App is running (^_^)`);
